Splits error.rs into separate files
