+++ full-service/Cargo.toml
@@ -41,6 +41,7 @@ displaydoc = {version = "0.1", default-features = false }
 dotenv = "0.15.0"
 grpcio = "0.6.0"
 hex = {version = "0.4", default-features = false }
+mockall = "0.9.0"
 num_cpus = "1.12"
 prost = { version = "0.6.1", default-features = false, features = ["prost-derive"] }
 rand = { version = "0.7", default-features = false }
