+++ full-service/src/db/wallet_db_error.rs
@@ -1,0 +1,112 @@
+use displaydoc::Display;
+
+#[derive(Display, Debug)]
+pub enum WalletDbError {
+    /// Diesel Error: {0}
+    Diesel(diesel::result::Error),
+
+    /// Error with rocket databases: {0}
+    RocketDB(rocket_contrib::databases::r2d2::Error),
+
+    /// Duplicate entries with the same ID {0}
+    DuplicateEntries(String),
+
+    /// Error encoding b58 {0}
+    B58Encode(mc_api::display::Error),
+
+    /// Error decoding b58: No public address in wrapper.
+    B58Decode,
+
+    /// Constructed a malformed transaction with multiple account IDs
+    MultipleAccountIDsInTransaction,
+
+    /// Constructed a transaction with multiple recipients (not currently supported for transaction logs)
+    MultipleRecipientsInTransaction,
+
+    /// Constructed a transaction with no recipient
+    TransactionLacksRecipient,
+
+    /// Constructed a transaction that is not linked to any account in the wallet
+    TransactionLacksAccount,
+
+    /// Error decoding prost {0}
+    ProstDecode(prost::DecodeError),
+
+    /// We expect one change output per TxProposal
+    UnexpectedNumberOfChangeOutputs,
+
+    /// Key Image missing when recovering orphaned Txo
+    MissingKeyImage,
+
+    /// Subaddress on received transaction is null
+    NullSubaddressOnReceived,
+
+    /// No unspent Txos in the wallet
+    NoSpendableTxos,
+
+    /// Txos are too fragmented to construct a transaction with MAX_INPUTS. Please combine txos.
+    InsufficientFundsFragmentedTxos,
+
+    /// Insufficient Funds {0}
+    InsufficientFunds(String),
+
+    /// Insufficient funds from Txos under max_spendable_value {0}
+    InsufficientFundsUnderMaxSpendable(String),
+
+    /// Multiple AccountTxoStatus entries for Txo
+    MultipleStatusesForTxo,
+
+    /// Unexpected TXO Type {0}
+    UnexpectedTransactionTxoType(String),
+
+    /// Transaction mismatch when retrieving associated Txos
+    TransactionMismatch,
+
+    /// Account Not Found {0}
+    AccountNotFound(String),
+
+    /// AssignedSubaddress Not Found {0}
+    AssignedSubaddressNotFound(String),
+
+    /// Txo Not Found {0}
+    TxoNotFound(String),
+
+    /// TransactionLog Not Found {0}
+    TransactionLogNotFound(String),
+
+    /// AccountTxoStatus not found {0}
+    AccountTxoStatusNotFound(String),
+
+    /// Cannot log a transaction with a value > i64::MAX
+    TransactionValueExceedsMax,
+
+    /// The Txo Exists, but for another account {0}
+    TxoExistsForAnotherAccount(String),
+
+    /// The Txo is associated with too many Accounts {0}
+    TxoAssociatedWithTooManyAccounts(String),
+}
+
+impl From<diesel::result::Error> for WalletDbError {
+    fn from(src: diesel::result::Error) -> Self {
+        Self::Diesel(src)
+    }
+}
+
+impl From<rocket_contrib::databases::r2d2::Error> for WalletDbError {
+    fn from(src: rocket_contrib::databases::r2d2::Error) -> Self {
+        Self::RocketDB(src)
+    }
+}
+
+impl From<mc_api::display::Error> for WalletDbError {
+    fn from(src: mc_api::display::Error) -> Self {
+        Self::B58Encode(src)
+    }
+}
+
+impl From<prost::DecodeError> for WalletDbError {
+    fn from(src: prost::DecodeError) -> Self {
+        Self::ProstDecode(src)
+    }
+}
