+++ full-service/src/service/mod.rs
@@ -4,11 +4,15 @@
 
 mod decorated_types;
 pub mod sync;
+pub mod sync_error;
 pub mod transaction_builder;
+pub mod transaction_builder_error;
 mod wallet_trait;
 mod wallet_api;
 mod wallet_service;
+mod wallet_service_error;
 
 pub use decorated_types::*;
 pub use wallet_api::{rocket, WalletApiState};
 pub use wallet_service::WalletService;
+pub use wallet_service_error::WalletServiceError;
