+++ full-service/src/service/wallet_trait.rs
@@ -1,9 +1,14 @@
 use crate::service::decorated_types::{JsonWalletStatus, JsonCreateAccountResponse, JsonAccount, JsonTxo, JsonBalanceResponse, JsonAddress, JsonSubmitResponse, JsonTransactionLog, JsonBlock, JsonBlockContents, JsonProof};
-use crate::error::WalletServiceError;
+use crate::service::WalletServiceError;
 use crate::db::account::AccountID;
 use mc_mobilecoind_json::data_types::{JsonTxProposal, JsonTx, JsonTxOut};
 
+#[cfg(test)]
+use mockall::*;
+
+
 /// A MobileCoin wallet.
+#[cfg_attr(test, automock)]
 pub trait Wallet {
 
     /// An overview of this wallet.
@@ -92,7 +97,7 @@ pub trait Wallet {
     fn create_assigned_subaddress(
         &self,
         account_id_hex: &str,
-        comment: Option<&str>,
+        comment: Option<String>,
         // FIXME: WS-32 - add "sync from block"
     ) -> Result<JsonAddress, WalletServiceError>;
 
@@ -121,7 +126,7 @@ pub trait Wallet {
         account_id_hex: &str,
         recipient_public_address: &str,
         value: String,
-        input_txo_ids: Option<&Vec<String>>,
+        input_txo_ids: Option<Vec<String>>,
         fee: Option<String>,
         tombstone_block: Option<String>,
         max_spendable_value: Option<String>,
@@ -157,7 +162,7 @@ pub trait Wallet {
         account_id_hex: &str,
         recipient_public_address: &str,
         value: String,
-        input_txo_ids: Option<&Vec<String>>,
+        input_txo_ids: Option<Vec<String>>,
         fee: Option<String>,
         tombstone_block: Option<String>,
         max_spendable_value: Option<String>,
@@ -227,4 +232,36 @@ pub trait Wallet {
         txo_id_hex: &str,
         proof_hex: &str,
     ) -> Result<bool, WalletServiceError>;
+}
+
+#[cfg(test)]
+mod tests {
+
+    use crate::service::wallet_trait::MockWallet;
+    use crate::service::JsonWalletStatus;
+
+    // /// Example of creating a mock Wallet
+    // #[test]
+    // fn mock_wallet_example() {
+    //     let mut mock_wallet = MockWallet::new();
+    //
+    //     // wallet status, now with more Pinnipeds.
+    //     let expected_status = JsonWalletStatus{
+    //         object: "LeopardSeal".to_string(),
+    //         network_height: "GreySeal".to_string(),
+    //         local_height: "HarbourSeal".to_string(),
+    //         is_synced_all: false,
+    //         total_available_pmob: "RibbonSeal".to_string(),
+    //         total_pending_pmob: "BeardedSeal".to_string(),
+    //         account_ids: vec![],
+    //         account_map: Default::default()
+    //     };
+    //
+    //     // get_wallet_status should be called once, and should return `expected_status`.
+    //     mock_wallet.expect_get_wallet_status()
+    //         .times(1)
+    //         .return_const(Ok(expected_status.clone()));
+    //
+    //     assert_eq!(mock_wallet.get_wallet_status(), Ok(expected_status));
+    // }
 }
